+++ README.md
@@ -68,5 +68,11 @@ simple example (toy psi)
 ```sh
 bazel run -c opt //mcpsi/example:toy_psi # run toy psi // PoC
 bazel run -c opt //mcpsi/example:toy_mc_psi # run toy circuit psi (sum) // PoC
-bazel run -c opt //mcpsi/example:mc_psi -- --set0 size_of_set0 --set1 size_of_set1 --interset size_of_interset # run circuit psi
+bazel run -c opt //mcpsi/example:mc_psi -- --set0 size_of_set0 --set1 size_of_set1 --interset size_of_interset --offline 0/1 # run malicious circuit psi (offline=0 for real cr and offline=1 for fake cr)
+```
+
+mcpsi under socket network
+```sh
+bazel run -c opt //mcpsi/example:mc_psi -- --set0 size_of_set0 --set1 size_of_set1 --interset size_of_interset --offline 0/1 --mode 1 --rank 0 # run malicious circuit psi for party 0
+bazel run -c opt //mcpsi/example:mc_psi -- --set0 size_of_set0 --set1 size_of_set1 --interset size_of_interset --offline 0/1 --mode 1 --rank 1 # run malicious circuit psi for party 1
 ```
