+++ mcpsi/cr/fake_cr.cc
@@ -4,9 +4,6 @@
 
 namespace mcpsi {
 
-// register string
-const std::string FakeCorrelation::id = std::string("FakeCorrelation");
-
 void FakeCorrelation::BeaverTriple(absl::Span<internal::ATy> a,
                                    absl::Span<internal::ATy> b,
                                    absl::Span<internal::ATy> c) {
@@ -38,6 +35,46 @@ void FakeCorrelation::BeaverTriple(absl::Span<internal::ATy> a,
     internal::Pack(absl::MakeConstSpan(b1), absl::MakeConstSpan(b_mac), b);
     internal::Pack(absl::MakeConstSpan(c1), absl::MakeConstSpan(c_mac), c);
   }
+}
+
+void FakeCorrelation::AuthSet(absl::Span<const internal::PTy> in,
+                              absl::Span<internal::ATy> out) {
+  RandomSet(out);
+  auto [val, mac] = internal::Unpack(out);
+  // val = in - val
+  internal::op::Sub(absl::MakeConstSpan(in), absl::MakeConstSpan(val),
+                    absl::MakeSpan(val));
+
+  auto conn = ctx_->GetConnection();
+  conn->SendAsync(
+      conn->NextRank(),
+      yacl::ByteContainerView(val.data(), val.size() * sizeof(internal::PTy)),
+      "AuthSet");
+
+  internal::op::Add(
+      absl::MakeConstSpan(mac),
+      absl::MakeConstSpan(internal::op::ScalarMul(key_, absl::MakeSpan(val))),
+      absl::MakeSpan(mac));
+  auto ret = internal::Pack(in, mac);
+  memcpy(out.data(), ret.data(), out.size() * sizeof(internal::ATy));
+}
+
+void FakeCorrelation::AuthGet(absl::Span<internal::ATy> out) {
+  RandomGet(out);
+  // val = 0
+  auto [val, mac] = internal::Unpack(out);
+
+  auto conn = ctx_->GetConnection();
+  auto recv_buf = conn->Recv(conn->NextRank(), "AuthSet");
+
+  auto diff = absl::MakeSpan(reinterpret_cast<internal::PTy*>(recv_buf.data()),
+                             out.size());
+
+  internal::op::Add(absl::MakeConstSpan(mac),
+                    absl::MakeConstSpan(internal::op::ScalarMul(key_, diff)),
+                    absl::MakeSpan(mac));
+  auto ret = internal::Pack(val, mac);
+  memcpy(out.data(), ret.data(), ret.size() * sizeof(internal::ATy));
 }
 
 void FakeCorrelation::RandomSet(absl::Span<internal::ATy> out) {
